feat: fixes to run it in modal.com
tbh this script was created a few months ago and I forgot most of the
reasons, but here is a summary of what I remember:

- Mount /models: here we save the Flux dev models, so the app/container
doesn't have to download the models everytime that is executed. And I
did it because sometimes the download from huggingface was failing so
it was not reliable.

- Mount /trainings: In this volume we'll save the models trained,
samples, and config files

- Update the base image: We use nvidia image because I was seeing this
error "OSError: CUDA_HOME environment variable is not set. Please set it
 to your CUDA install root"